Move function to generate doc into an markdown_formatter
